Refactor plan retrieval endpoints and update API usage
Moved plan retrieval endpoints from app_kernel.py to v3/api/router.py and refactored them for v3 API. Updated DataType enum values for consistency. Adjusted frontend API service to use new v3 endpoints and removed unused agent message and message methods.
