+++ src/backend/app_kernel.py
@@ -2,6 +2,7 @@
 import asyncio
 import logging
 import os
+
 # Azure monitoring
 import re
 import uuid
@@ -12,22 +13,40 @@ from auth.auth_utils import get_authenticated_user_details
 from azure.monitor.opentelemetry import configure_azure_monitor
 from common.config.app_config import config
 from common.database.database_factory import DatabaseFactory
-from common.models.messages_kernel import (AgentMessage, AgentType,
-                                           HumanClarification, HumanFeedback,
-                                           InputTask, Plan, PlanStatus,
-                                           PlanWithSteps, Step, UserLanguage)
+from common.models.messages_kernel import (
+    AgentMessage,
+    AgentType,
+    HumanClarification,
+    HumanFeedback,
+    InputTask,
+    Plan,
+    PlanStatus,
+    PlanWithSteps,
+    Step,
+    UserLanguage,
+)
 from common.utils.event_utils import track_event_if_configured
 from common.utils.utils_date import format_dates_in_messages
+
 # Updated import for KernelArguments
 from common.utils.utils_kernel import rai_success
+
 # FastAPI imports
-from fastapi import (FastAPI, HTTPException, Query, Request, WebSocket,
-                     WebSocketDisconnect)
+from fastapi import (
+    FastAPI,
+    HTTPException,
+    Query,
+    Request,
+    WebSocket,
+    WebSocketDisconnect,
+)
 from fastapi.middleware.cors import CORSMiddleware
 from kernel_agents.agent_factory import AgentFactory
+
 # Local imports
 from middleware.health_check import HealthCheckMiddleware
 from v3.api.router import app_v3
+
 # Semantic Kernel imports
 from v3.orchestration.orchestration_manager import OrchestrationManager
 
@@ -110,841 +129,6 @@ async def user_browser_language_endpoint(user_language: UserLanguage, request: R
     logging.info(f"Received browser language '{user_language}' for user ")
 
     return {"status": "Language received successfully"}
-
-
-# @app.post("/api/input_task")
-# async def input_task_endpoint(input_task: InputTask, request: Request):
-#     """
-#     Receive the initial input task from the user.
-#     """
-#     # Fix 1: Properly await the async rai_success function
-#     if not await rai_success(input_task.description, True):
-#         print("RAI failed")
-
-#         track_event_if_configured(
-#             "RAI failed",
-#             {
-#                 "status": "Plan not created - RAI validation failed",
-#                 "description": input_task.description,
-#                 "session_id": input_task.session_id,
-#             },
-#         )
-
-#         return {
-#             "status": "RAI_VALIDATION_FAILED",
-#             "message": "Content Safety Check Failed",
-#             "detail": "Your request contains content that doesn't meet our safety guidelines. Please modify your request to ensure it's appropriate and try again.",
-#             "suggestions": [
-#                 "Remove any potentially harmful, inappropriate, or unsafe content",
-#                 "Use more professional and constructive language",
-#                 "Focus on legitimate business or educational objectives",
-#                 "Ensure your request complies with content policies",
-#             ],
-#         }
-#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-#     user_id = authenticated_user["user_principal_id"]
-
-#     if not user_id:
-#         track_event_if_configured(
-#             "UserIdNotFound", {"status_code": 400, "detail": "no user"}
-#         )
-#         raise HTTPException(status_code=400, detail="no user")
-
-#     # Generate session ID if not provided
-#     if not input_task.session_id:
-#         input_task.session_id = str(uuid.uuid4())
-
-#     try:
-#         # Create all agents instead of just the planner agent
-#         # This ensures other agents are created first and the planner has access to them
-#         memory_store = await DatabaseFactory.get_database(user_id=user_id)
-#         client = None
-#         try:
-#             client = config.get_ai_project_client()
-#         except Exception as client_exc:
-#             logging.error(f"Error creating AIProjectClient: {client_exc}")
-
-#         agents = await AgentFactory.create_all_agents(
-#             session_id=input_task.session_id,
-#             user_id=user_id,
-#             memory_store=memory_store,
-#             client=client,
-#         )
-
-#         group_chat_manager = agents[AgentType.GROUP_CHAT_MANAGER.value]
-
-#         # Convert input task to JSON for the kernel function, add user_id here
-
-#         # Use the planner to handle the task
-#         await group_chat_manager.handle_input_task(input_task)
-
-#         # Get plan from memory store
-#         plan = await memory_store.get_plan_by_session(input_task.session_id)
-
-#         if not plan:  # If the plan is not found, raise an error
-#             track_event_if_configured(
-#                 "PlanNotFound",
-#                 {
-#                     "status": "Plan not found",
-#                     "session_id": input_task.session_id,
-#                     "description": input_task.description,
-#                 },
-#             )
-#             raise HTTPException(status_code=404, detail="Plan not found")
-#         # Log custom event for successful input task processing
-#         track_event_if_configured(
-#             "InputTaskProcessed",
-#             {
-#                 "status": f"Plan created with ID: {plan.id}",
-#                 "session_id": input_task.session_id,
-#                 "plan_id": plan.id,
-#                 "description": input_task.description,
-#             },
-#         )
-#         if client:
-#             try:
-#                 client.close()
-#             except Exception as e:
-#                 logging.error(f"Error sending to AIProjectClient: {e}")
-#         return {
-#             "status": f"Plan created with ID: {plan.id}",
-#             "session_id": input_task.session_id,
-#             "plan_id": plan.id,
-#             "description": input_task.description,
-#         }
-
-#     except Exception as e:
-#         # Extract clean error message for rate limit errors
-#         error_msg = str(e)
-#         if "Rate limit is exceeded" in error_msg:
-#             match = re.search(
-#                 r"Rate limit is exceeded\. Try again in (\d+) seconds?\.", error_msg
-#             )
-#             if match:
-#                 error_msg = "Application temporarily unavailable due to quota limits. Please try again later."
-
-#         track_event_if_configured(
-#             "InputTaskError",
-#             {
-#                 "session_id": input_task.session_id,
-#                 "description": input_task.description,
-#                 "error": str(e),
-#             },
-#         )
-#         raise HTTPException(
-#             status_code=400, detail=f"Error creating plan: {error_msg}"
-#         ) from e
-
-
-@app.post("/api/human_feedback")
-async def human_feedback_endpoint(human_feedback: HumanFeedback, request: Request):
-    """
-    Receive human feedback on a step.
-
-    ---
-    tags:
-      - Feedback
-    parameters:
-      - name: user_principal_id
-        in: header
-        type: string
-        required: true
-        description: User ID extracted from the authentication header
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            step_id:
-              type: string
-              description: The ID of the step to provide feedback for
-            plan_id:
-              type: string
-              description: The plan ID
-            session_id:
-              type: string
-              description: The session ID
-            approved:
-              type: boolean
-              description: Whether the step is approved
-            human_feedback:
-              type: string
-              description: Optional feedback details
-            updated_action:
-              type: string
-              description: Optional updated action
-            user_id:
-              type: string
-              description: The user ID providing the feedback
-    responses:
-      200:
-        description: Feedback received successfully
-        schema:
-          type: object
-          properties:
-            status:
-              type: string
-            session_id:
-              type: string
-            step_id:
-              type: string
-      400:
-        description: Missing or invalid user information
-    """
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-    if not user_id:
-        track_event_if_configured(
-            "UserIdNotFound", {"status_code": 400, "detail": "no user"}
-        )
-        raise HTTPException(status_code=400, detail="no user")
-
-    memory_store = await DatabaseFactory.get_database(user_id=user_id)
-
-    client = None
-    try:
-        client = config.get_ai_project_client()
-    except Exception as client_exc:
-        logging.error(f"Error creating AIProjectClient: {client_exc}")
-
-    human_agent = await AgentFactory.create_agent(
-        agent_type=AgentType.HUMAN,
-        session_id=human_feedback.session_id,
-        user_id=user_id,
-        memory_store=memory_store,
-        client=client,
-    )
-
-    if human_agent is None:
-        track_event_if_configured(
-            "AgentNotFound",
-            {
-                "status": "Agent not found",
-                "session_id": human_feedback.session_id,
-                "step_id": human_feedback.step_id,
-            },
-        )
-        raise HTTPException(status_code=404, detail="Agent not found")
-
-    # Use the human agent to handle the feedback
-    await human_agent.handle_human_feedback(human_feedback=human_feedback)
-
-    track_event_if_configured(
-        "Completed Feedback received",
-        {
-            "status": "Feedback received",
-            "session_id": human_feedback.session_id,
-            "step_id": human_feedback.step_id,
-        },
-    )
-    if client:
-        try:
-            client.close()
-        except Exception as e:
-            logging.error(f"Error sending to AIProjectClient: {e}")
-    return {
-        "status": "Feedback received",
-        "session_id": human_feedback.session_id,
-        "step_id": human_feedback.step_id,
-    }
-
-
-@app.post("/api/human_clarification_on_plan")
-async def human_clarification_endpoint(
-    human_clarification: HumanClarification, request: Request
-):
-    """
-    Receive human clarification on a plan.
-
-    ---
-    tags:
-      - Clarification
-    parameters:
-      - name: user_principal_id
-        in: header
-        type: string
-        required: true
-        description: User ID extracted from the authentication header
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            plan_id:
-              type: string
-              description: The plan ID requiring clarification
-            session_id:
-              type: string
-              description: The session ID
-            human_clarification:
-              type: string
-              description: Clarification details provided by the user
-            user_id:
-              type: string
-              description: The user ID providing the clarification
-    responses:
-      200:
-        description: Clarification received successfully
-        schema:
-          type: object
-          properties:
-            status:
-              type: string
-            session_id:
-              type: string
-      400:
-        description: Missing or invalid user information
-    """
-    if not await rai_success(human_clarification.human_clarification, False):
-        print("RAI failed")
-        track_event_if_configured(
-            "RAI failed",
-            {
-                "status": "Clarification rejected - RAI validation failed",
-                "description": human_clarification.human_clarification,
-                "session_id": human_clarification.session_id,
-            },
-        )
-        raise HTTPException(
-            status_code=400,
-            detail={
-                "error_type": "RAI_VALIDATION_FAILED",
-                "message": "Clarification Safety Check Failed",
-                "description": "Your clarification contains content that doesn't meet our safety guidelines. Please provide a more appropriate clarification.",
-                "suggestions": [
-                    "Use clear and professional language",
-                    "Avoid potentially harmful or inappropriate content",
-                    "Focus on providing constructive feedback or clarification",
-                    "Ensure your message complies with content policies",
-                ],
-                "user_action": "Please revise your clarification and try again",
-            },
-        )
-
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-    if not user_id:
-        track_event_if_configured(
-            "UserIdNotFound", {"status_code": 400, "detail": "no user"}
-        )
-        raise HTTPException(status_code=400, detail="no user")
-
-    memory_store = await DatabaseFactory.get_database(user_id=user_id)
-    client = None
-    try:
-        client = config.get_ai_project_client()
-    except Exception as client_exc:
-        logging.error(f"Error creating AIProjectClient: {client_exc}")
-
-    human_agent = await AgentFactory.create_agent(
-        agent_type=AgentType.HUMAN,
-        session_id=human_clarification.session_id,
-        user_id=user_id,
-        memory_store=memory_store,
-        client=client,
-    )
-
-    if human_agent is None:
-        track_event_if_configured(
-            "AgentNotFound",
-            {
-                "status": "Agent not found",
-                "session_id": human_clarification.session_id,
-                "step_id": human_clarification.step_id,
-            },
-        )
-        raise HTTPException(status_code=404, detail="Agent not found")
-
-    # Use the human agent to handle the feedback
-    await human_agent.handle_human_clarification(
-        human_clarification=human_clarification
-    )
-
-    track_event_if_configured(
-        "Completed Human clarification on the plan",
-        {
-            "status": "Clarification received",
-            "session_id": human_clarification.session_id,
-        },
-    )
-    if client:
-        try:
-            client.close()
-        except Exception as e:
-            logging.error(f"Error sending to AIProjectClient: {e}")
-    return {
-        "status": "Clarification received",
-        "session_id": human_clarification.session_id,
-    }
-
-
-@app.post("/api/approve_step_or_steps")
-async def approve_step_endpoint(
-    human_feedback: HumanFeedback, request: Request
-) -> Dict[str, str]:
-    """
-    Approve a step or multiple steps in a plan.
-
-    ---
-    tags:
-      - Approval
-    parameters:
-      - name: user_principal_id
-        in: header
-        type: string
-        required: true
-        description: User ID extracted from the authentication header
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            step_id:
-              type: string
-              description: Optional step ID to approve
-            plan_id:
-              type: string
-              description: The plan ID
-            session_id:
-              type: string
-              description: The session ID
-            approved:
-              type: boolean
-              description: Whether the step(s) are approved
-            human_feedback:
-              type: string
-              description: Optional feedback details
-            updated_action:
-              type: string
-              description: Optional updated action
-            user_id:
-              type: string
-              description: The user ID providing the approval
-    responses:
-      200:
-        description: Approval status returned
-        schema:
-          type: object
-          properties:
-            status:
-              type: string
-      400:
-        description: Missing or invalid user information
-    """
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-    if not user_id:
-        track_event_if_configured(
-            "UserIdNotFound", {"status_code": 400, "detail": "no user"}
-        )
-        raise HTTPException(status_code=400, detail="no user")
-
-    # Get the agents for this session
-    memory_store = await DatabaseFactory.get_database(user_id=user_id)
-    client = None
-    try:
-        client = config.get_ai_project_client()
-    except Exception as client_exc:
-        logging.error(f"Error creating AIProjectClient: {client_exc}")
-    agents = await AgentFactory.create_all_agents(
-        session_id=human_feedback.session_id,
-        user_id=user_id,
-        memory_store=memory_store,
-        client=client,
-    )
-
-    # Send the approval to the group chat manager
-    group_chat_manager = agents[AgentType.GROUP_CHAT_MANAGER.value]
-
-    await group_chat_manager.handle_human_feedback(human_feedback)
-
-    if client:
-        try:
-            client.close()
-        except Exception as e:
-            logging.error(f"Error sending to AIProjectClient: {e}")
-    # Return a status message
-    if human_feedback.step_id:
-        track_event_if_configured(
-            "Completed Human clarification with step_id",
-            {
-                "status": f"Step {human_feedback.step_id} - Approval:{human_feedback.approved}."
-            },
-        )
-
-        return {
-            "status": f"Step {human_feedback.step_id} - Approval:{human_feedback.approved}."
-        }
-    else:
-        track_event_if_configured(
-            "Completed Human clarification without step_id",
-            {"status": "All steps approved"},
-        )
-
-        return {"status": "All steps approved"}
-
-
-# Get plans is called in the initial side rendering of the frontend
-@app.get("/api/plans")
-async def get_plans(
-    request: Request,
-    plan_id: Optional[str] = Query(None),
-):
-    """
-    Retrieve plans for the current user.
-
-    ---
-    tags:
-      - Plans
-    parameters:
-      - name: session_id
-        in: query
-        type: string
-        required: false
-        description: Optional session ID to retrieve plans for a specific session
-    responses:
-      200:
-        description: List of plans with steps for the user
-        schema:
-          type: array
-          items:
-            type: object
-            properties:
-              id:
-                type: string
-                description: Unique ID of the plan
-              session_id:
-                type: string
-                description: Session ID associated with the plan
-              initial_goal:
-                type: string
-                description: The initial goal derived from the user's input
-              overall_status:
-                type: string
-                description: Status of the plan (e.g., in_progress, completed)
-              steps:
-                type: array
-                items:
-                  type: object
-                  properties:
-                    id:
-                      type: string
-                      description: Unique ID of the step
-                    plan_id:
-                      type: string
-                      description: ID of the plan the step belongs to
-                    action:
-                      type: string
-                      description: The action to be performed
-                    agent:
-                      type: string
-                      description: The agent responsible for the step
-                    status:
-                      type: string
-                      description: Status of the step (e.g., planned, approved, completed)
-      400:
-        description: Missing or invalid user information
-      404:
-        description: Plan not found
-    """
-
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-    if not user_id:
-        track_event_if_configured(
-            "UserIdNotFound", {"status_code": 400, "detail": "no user"}
-        )
-        raise HTTPException(status_code=400, detail="no user")
-
-    #### <To do: Francia> Replace the following with code to get plan run history from the database
-
-    # # Initialize memory context
-    memory_store = await DatabaseFactory.get_database(user_id=user_id)
-
-    if plan_id:
-        plan = await memory_store.get_plan_by_plan_id(plan_id=plan_id)
-        if not plan:
-            track_event_if_configured(
-                "GetPlanBySessionNotFound",
-                {"status_code": 400, "detail": "Plan not found"},
-            )
-            raise HTTPException(status_code=404, detail="Plan not found")
-
-        # Use get_steps_by_plan to match the original implementation
-        steps = await memory_store.get_steps_by_plan(plan_id=plan.id)
-        messages = await memory_store.get_data_by_type_and_session_id(
-            "agent_message", session_id=plan.session_id
-        )
-
-        plan_with_steps = PlanWithSteps(**plan.model_dump(), steps=steps)
-        plan_with_steps.update_step_counts()
-
-        # Format dates in messages according to locale
-        formatted_messages = format_dates_in_messages(
-            messages, config.get_user_local_browser_language()
-        )
-
-        return [plan_with_steps, formatted_messages]
-
-    current_team = await memory_store.get_current_team(user_id=user_id)
-    if not current_team:
-        return []
-
-    all_plans = await memory_store.get_all_plans_by_team_id(team_id=current_team.id)
-    # Fetch steps for all plans concurrently
-    steps_for_all_plans = await asyncio.gather(
-        *[memory_store.get_steps_by_plan(plan_id=plan.id) for plan in all_plans]
-    )
-    # Create list of PlanWithSteps and update step counts
-    list_of_plans_with_steps = []
-    for plan, steps in zip(all_plans, steps_for_all_plans):
-        plan_with_steps = PlanWithSteps(**plan.model_dump(), steps=steps)
-        plan_with_steps.update_step_counts()
-        list_of_plans_with_steps.append(plan_with_steps)
-
-    return []
-
-
-@app.get("/api/steps/{plan_id}", response_model=List[Step])
-async def get_steps_by_plan(plan_id: str, request: Request) -> List[Step]:
-    """
-    Retrieve steps for a specific plan.
-
-    ---
-    tags:
-      - Steps
-    parameters:
-      - name: plan_id
-        in: path
-        type: string
-        required: true
-        description: The ID of the plan to retrieve steps for
-    responses:
-      200:
-        description: List of steps associated with the specified plan
-        schema:
-          type: array
-          items:
-            type: object
-            properties:
-              id:
-                type: string
-                description: Unique ID of the step
-              plan_id:
-                type: string
-                description: ID of the plan the step belongs to
-              action:
-                type: string
-                description: The action to be performed
-              agent:
-                type: string
-                description: The agent responsible for the step
-              status:
-                type: string
-                description: Status of the step (e.g., planned, approved, completed)
-              agent_reply:
-                type: string
-                description: Optional response from the agent after execution
-              human_feedback:
-                type: string
-                description: Optional feedback provided by a human
-              updated_action:
-                type: string
-                description: Optional modified action based on feedback
-       400:
-        description: Missing or invalid user information
-      404:
-        description: Plan or steps not found
-    """
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-    if not user_id:
-        track_event_if_configured(
-            "UserIdNotFound", {"status_code": 400, "detail": "no user"}
-        )
-        raise HTTPException(status_code=400, detail="no user")
-
-    # Initialize memory context
-    memory_store = await DatabaseFactory.get_database(user_id=user_id)
-    steps = await memory_store.get_steps_for_plan(plan_id=plan_id)
-    return steps
-
-
-@app.get("/api/agent_messages/{session_id}", response_model=List[AgentMessage])
-async def get_agent_messages(session_id: str, request: Request) -> List[AgentMessage]:
-    """
-    Retrieve agent messages for a specific session.
-
-    ---
-    tags:
-      - Agent Messages
-    parameters:
-      - name: session_id
-        in: path
-        type: string
-        required: true
-        in: path
-        type: string
-        required: true
-        description: The ID of the session to retrieve agent messages for
-    responses:
-      200:
-        description: List of agent messages associated with the specified session
-        schema:
-          type: array
-          items:
-            type: object
-            properties:
-              id:
-                type: string
-                description: Unique ID of the agent message
-              session_id:
-                type: string
-                description: Session ID associated with the message
-              plan_id:
-                type: string
-                description: Plan ID related to the agent message
-              content:
-                type: string
-                description: Content of the message
-              source:
-                type: string
-                description: Source of the message (e.g., agent type)
-              timestamp:
-                type: string
-                format: date-time
-                description: Timestamp of the message
-              step_id:
-                type: string
-                description: Optional step ID associated with the message
-      400:
-        description: Missing or invalid user information
-      404:
-        description: Agent messages not found
-    """
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-    if not user_id:
-        track_event_if_configured(
-            "UserIdNotFound", {"status_code": 400, "detail": "no user"}
-        )
-        raise HTTPException(status_code=400, detail="no user")
-
-    # Initialize memory context
-    memory_store = await DatabaseFactory.get_database(user_id=user_id)
-    agent_messages = await memory_store.get_data_by_type("agent_message")
-    return agent_messages
-
-
-@app.get("/api/agent-tools")
-async def get_agent_tools():
-    """
-    Retrieve all available agent tools.
-
-    ---
-    tags:
-      - Agent Tools
-    responses:
-      200:
-        description: List of all available agent tools and their descriptions
-        schema:
-          type: array
-          items:
-            type: object
-            properties:
-              agent:
-                type: string
-                description: Name of the agent associated with the tool
-              function:
-                type: string
-                description: Name of the tool function
-              description:
-                type: string
-                description: Detailed description of what the tool does
-              arguments:
-                type: string
-                description: Arguments required by the tool function
-    """
-    return []
-
-
-# @app.post("/api/test/streaming/{plan_id}")
-# async def test_streaming_updates(plan_id: str):
-#     """
-#     Test endpoint to simulate streaming updates for a plan.
-#     This is for testing the WebSocket streaming functionality.
-#     """
-#     from common.utils.websocket_streaming import (send_agent_message,
-#                                                   send_plan_update,
-#                                                   send_step_update)
-
-#     try:
-#         # Simulate a series of streaming updates
-#         await send_agent_message(
-#             plan_id=plan_id,
-#             agent_name="Data Analyst",
-#             content="Starting analysis of the data...",
-#             message_type="thinking",
-#         )
-
-#         await asyncio.sleep(1)
-
-#         await send_plan_update(
-#             plan_id=plan_id,
-#             step_id="step_1",
-#             agent_name="Data Analyst",
-#             content="Analyzing customer data patterns...",
-#             status="in_progress",
-#             message_type="action",
-#         )
-
-#         await asyncio.sleep(2)
-
-#         await send_agent_message(
-#             plan_id=plan_id,
-#             agent_name="Data Analyst",
-#             content="Found 3 key insights in the customer data. Processing recommendations...",
-#             message_type="result",
-#         )
-
-#         await asyncio.sleep(1)
-
-#         await send_step_update(
-#             plan_id=plan_id,
-#             step_id="step_1",
-#             status="completed",
-#             content="Data analysis completed successfully!",
-#         )
-
-#         await send_agent_message(
-#             plan_id=plan_id,
-#             agent_name="Business Advisor",
-#             content="Reviewing the analysis results and preparing strategic recommendations...",
-#             message_type="thinking",
-#         )
-
-#         await asyncio.sleep(2)
-
-#         await send_plan_update(
-#             plan_id=plan_id,
-#             step_id="step_2",
-#             agent_name="Business Advisor",
-#             content="Based on the data analysis, I recommend focusing on customer retention strategies for the identified high-value segments.",
-#             status="completed",
-#             message_type="result",
-#         )
-
-#         return {
-#             "status": "success",
-#             "message": f"Test streaming updates sent for plan {plan_id}",
-#         }
-
-#     except Exception as e:
-#         logging.error(f"Error sending test streaming updates: {e}")
-#         raise HTTPException(status_code=500, detail=str(e))
 
 
 # Run the app
